backend/requirements.txt

flask==3.0.2
flask-cors==4.0.0
firebase-admin==6.4.0
google-cloud-storage==2.14.0
gunicorn==21.2.0
werkzeug==3.0.1 
